refactor(comet-sdk): extract shared responses protocol package
